refactor and add FinalBid component
