+++ pages/gameboard/final/index.js
@@ -1,18 +1,12 @@
 import { useRouter } from "next/router";
 import { useGame } from "../../../context/GameContext";
 import PageContainer from "../../../components/PageContainer";
-import { useEffect, useState } from "react";
-import styles from "./styles.module.css";
+import { useEffect } from "react";
+import FinalBid from "../../../components/FinalBid";
 
 const FinalGameboard = () => {
-  const [bid, setBid] = useState(1);
-  const [isBidValid, setIsBidValid] = useState(true);
-
-  const { finalSlug, finalCategory, playerScores, setAttempts } = useGame();
+  const { finalSlug } = useGame();
   const router = useRouter();
-
-  const playerScore = playerScores.player1.score;
-  const playerScoreString = playerScore.toString();
 
   useEffect(() => {
     if (!finalSlug) {
@@ -32,51 +26,9 @@ const FinalGameboard = () => {
     };
   }, []);
 
-  const handleSubmit = () => {
-    if (bid > playerScore || bid < 0) {
-      setIsBidValid(false);
-    } else {
-      setAttempts(["FINAL"]);
-      router.push({
-        pathname: `/gameboard/final/questions/${finalSlug}`,
-        query: { value: bid },
-      });
-    }
-  };
-
   return (
     <PageContainer>
-      <h1 className={styles.header}>
-        WELCOME TO FINAL <br /> JEOPARDY!
-      </h1>
-      <h1>
-        TODAY'S CATEGORY IS
-        <span className={styles.finalCategory}> {finalCategory}!</span>
-      </h1>
-      <h2 className={styles.bidMessage}>
-        Please Enter Your Bid Before Proceeding <br />
-        to the Final Question:
-      </h2>
-      <div className={styles.submitContainer}>
-        <span>$</span>
-        <input
-          id="bid"
-          type="number"
-          placeholder="Place bid here..."
-          value={bid}
-          onChange={(e) => setBid(e.target.value)}
-          autoComplete="off"
-          min="1"
-          max={playerScoreString}
-          onKeyDown={(e) => e.key === "Enter" && handleSubmit()}
-        />
-        <button onClick={handleSubmit}>Submit</button>
-      </div>
-      {!isBidValid && (
-        <span className={styles.validation}>
-          {` Bid must be greater than 0 and no more than your current score {$${playerScore}}!`}
-        </span>
-      )}
+      <FinalBid />
     </PageContainer>
   );
 };
